requirements.txt
streamlit==1.45.1
streamlit-webrtc==0.62.4
elevenlabs==2.3.0
python-dotenv==1.1.0
twilio==9.6.3
SpeechRecognition==3.14.3
webrtcvad-wheels==2.0.14
streamlit-realtime-audio-recorder==0.1.2